Move back `SpatioTemporalProblem` to its own module
