+++ src/moscot/problems/space/__init__.py
@@ -1,12 +1,10 @@
 from ._alignment import AlignmentProblem
 from ._mapping import MappingProblem
 from ._mixins import SpatialAlignmentMixin, SpatialMappingMixin
-from ._spatio_temporal import SpatioTemporalProblem
 
 __all__ = [
     "AlignmentProblem",
     "MappingProblem",
-    "SpatioTemporalProblem",
     "SpatialMappingMixin",
     "SpatialAlignmentMixin",
 ]
