+++ src/moscot/problems/spatiotemporal/__init__.py
@@ -1,0 +1,3 @@
+from ._spatio_temporal import SpatioTemporalProblem
+
+__all__ = ["SpatioTemporalProblem"]
